pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "igor_portifolio_backend"
version = "0.1.0"
description = "Agente IA com LangGraph · LangChain · LangMem"
authors = [{name = "Igor Santos de Macedo"}]
readme = "README.md"
requires-python = ">=3.10"

# As mesmas libs do requirements.txt
dependencies = [
    "python-dotenv",
    "tiktoken",
    "pydantic",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-groq>=0.2.0",
    "langgraph>=0.2.0",
    "langmem",
    "chromadb>=0.4.0",
    "langchain-chroma",
    "langchain-huggingface",
    "sentence-transformers>=2.2.0"
]

[tool.setuptools.packages.find]
where   = ["."]
include = ["core*"]      # garante que 'core', 'core.*' entrem no pacote

[project.optional-dependencies]
dev = ["pytest", "ruff"]
